Migrate ScheduleCompetitionScoreUpdatesJob
