+++ server/src/api/jobs/instances/ScheduleCompetitionScoreUpdatesJob.ts
@@ -1,7 +1,5 @@
-import prisma from '../../../prisma';
-import { Period, PeriodProps } from '../../..//utils';
+import { ScheduleCompetitionScoreUpdatesJob as NewScheduleCompetitionScoreUpdatesJob } from '../../../jobs/instances/ScheduleCompetitionScoreUpdatesJob';
 import { JobType, JobDefinition } from '../job.types';
-import { jobManager } from '..';
 
 class ScheduleCompetitionScoreUpdatesJob implements JobDefinition<unknown> {
   type: JobType;
@@ -11,22 +9,10 @@ class ScheduleCompetitionScoreUpdatesJob implements JobDefinition<unknown> {
   }
 
   async execute() {
-    const allCompetitions = await prisma.competition.findMany({
-      where: {
-        OR: [{ endsAt: { gt: new Date() } }, { score: { gt: 0 } }]
-      },
-      select: { id: true }
-    });
-
-    // Distribute these evenly throughout the 12h, with a variable cooldown between each
-    const cooldown = Math.floor(PeriodProps[Period.DAY].milliseconds / 2 / allCompetitions.length);
-
-    allCompetitions.forEach((competition, i) => {
-      jobManager.add(
-        { type: JobType.UPDATE_COMPETITION_SCORE, payload: { competitionId: competition.id } },
-        { delay: i * cooldown }
-      );
-    });
+    // We're migrating to a new job manager, but jobs for the current one are still in-queue
+    // so to prevent duplicating code, just use the old job manager to execute the job on the new one
+    // Once this old job is no longer in use, we can remove this entire file.
+    await new NewScheduleCompetitionScoreUpdatesJob().execute();
   }
 }
